review_app/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
google-genai>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
